Basic repository setup (#7)
This repository was initialized from our template repo, which is
intended for libraries written in TypeScript primarily. This repository
will be a JavaScript library and CLI tool though, so much of this
initial configuration wasn't applicable.

Here's a summary of the changes made:
* README updated with package name and description
* `package.json` updated with package name and description
* ESLint config updated to reflect typical settings for a Node.js
  library that uses Jest for testing
* Test updated to be alongside code-under-test
* Remove build step from CI configuration
* Jest configuration updated to run tests using Node.js, not `ts-node`.
* Unused dependencies removed
* Add 'root: true' to ESLint config
* Test before publishing
* Update `@metamask/eslint-config` and related packages to v6

The actual auto-changelog code will be added in a later PR, along with
further documentation and tests.