Migrate original `auto-changelog.js` script (#8)
This script has been migrated from the `metamask-extension` repo [1].
Functionally it should behave almost identically. The differences are:

* It now gets the current version from the `npm_package_version`
environment variable [2] instead of from the hard-coded path to the
MetaMask extension manifest.
* The `CHANGELOG.md` file is now read from the current directory where
the script is run, rather than from the hard-coded path to the
`CHANGELOG.md` file in this repository.
* The repository URL is taken from the `npm_package_repository_url`
environment variable instead of being hard-coded.

These changes can all be seen in `src/cli.js`, which is the equivalent
to the top-level script in the `metamask-extension` repo,
`auto-changelog.js`.

These things will all be made more configurable in future PRs.

A mostly-empty `CHANGELOG.md` file has been added as well. It doesn't
quite work with the `update-changelog` script yet, because it wasn't
meant to accommodate packages with no releases. But it was enough to
make it not completely blow up.

One silly test has been added to allow CI to pass without removing
tests altogether. This will be improved in future PRs.

The package has one exported function, `updateChangelog`, to allow
the changelog to be updated programmatically.

[1]: https://github.com/MetaMask/metamask-extension/blob/0dbd678f393ae141d207e35b0132e2ee8cb118e9/development/auto-changelog.js
[2]: https://docs.npmjs.com/cli/v6/using-npm/scripts#packagejson-vars