Allow updating changelog when there are no tags
The changelog can now be updated even when there are no tags.
Previously it would break in `updateChangelog` when there was no most
recent tag, and if you got past that it would put `vundefined` in the
link definition for the "Unreleased" header. Now it is tolerant of
there being no tags, and the repo URL is used for the link definition
of the "Unreleased" header if there is no release to compare to.

Fixes #13