Allow configuring 'currentVersion'
Rather than always obtaining the 'currentVersion' of the project from
an environment variable, it can now be passed in via a CLI flag
instead.

Additionally, the validation for the version has been improved. It will
now throw if the current version isn't a valid SemVer version, as our
update logic depends upon it being valid. We also now ensure that the
error message for a messing version is only printed when it was
strictly required (it wasn't used for non-RC updates), and that the
script terminates at that point.

Closes #10