Fix release header sorting (#29)
The release header sorting was broken in two ways. First, the sort
function was totally broken - it returned a boolean instead of the
expected `-1` or `1`. Second, the release link generation would
always consider the _numerical_ previous release to be the previous
release for comparison purposes, even if it was released _after_.

The sorting function now correctly returns numbers, as
`Array.prototype.sort` expects. The release link logic has been updated
to ensure the "previous" release is both numerically and
chronologically previous as well.

We don't yet validate that the releases are sorted chronologically, but
that is the assumption for now. We can add validation using release
dates at some point in the future.