Allow configuring repository URL
The repository URL can now be configured via the `--repo` CLI flag, in
addition to being configurable via the `npm_package_repository_url`
environment variable.

Additional validation has been added to ensure the given URL is valid.
It now exits the process when the repo URL is invalid or isn't set, as
it wouldn't work properly in that case anyway.

Closes #12