Add `root` parameter
The `root` parameter allows specifying the root project directory. This
is used to resolve the changelog file path (unless an absolute path is
specified), and it's used by the `update` command to ensure the search
for changes since the last release only includes changes in the given
root directory.

This new option is exposed via the API only on the `update` function,
with the `rootProjectDirectory` option.

This was the last feature required for monorepo support.

Closes #34