Add unit tests for `parseChangelog` (#45)
`parseChangelog.js` now has 100% test coverage. This includes two unit
tests that demonstrate the undesired behaviour reported in #38 (the
last two).

These tests were added primarily to make it easier to fix #38 while
ensuring the desired behaviour is preserved.

One branch was removed from `parseChangelog` because it was
unreachable. Also a minor typo was fixed in a comment.