Allow multi-line changelog entries
Changelog entries are now allowed to span multiple lines and contain
sub-bullets. We even preserve trailing newlines in change descriptions,
because why not? Enforcing newlines isn't required to make sense of the
format.

Additional tests have been added to reflect the changes. A test has
been added to reflect that newlines aren't required between sections
in parsed changelogs either (this functionality was pre-existing, but
the absence was highlighted by the coverage report).

Fixes #38