Remove trailing '.git' from default repo URL (#74)
If no `repo` parameter is supplied we get the GitHub repository URL
from the environment variable `npm_package_repository_url`, which is
set to the git repository URL if the `repository` field of
`package.json` is set.

However, the git repository URL is not the same as the GitHub
repository URL. The git repository URL ends with `.git`, but the GitHub
repository URL does not. This discrepency wasn't noticed at first
because the `repository` field in the `package.json` file for this
project is malformed (the `.git` is missing).

The CLI has been updated to remove this extraneous `.git`, if it's
present. So how the CLI will work whether the trailing `.git` is there
or not. The `repository` entry for this project has been fixed as well.