Skip validation of "Uncategorized" header for older releases (#106)
Currently the `validate` command will ensure that no release has
uncategorized changes when the `--rc` flag is present, but it doesn't
care about uncategorized changes when the `--rc` flag isn't present.

This is against the spirit of the `--rc` flag. That flag should enable
extra validation specific to the preparation of the current release. It
shouldn't impose further burdens on maintainers that have nothing to do
with the current release. This validation is also very inconvenient for
some projects (e.g. `metamask-extension`), where the current
maintainers may not have all of the context to make improvements to
ancient changelog entries.