Support alternate tag prefixes (#120)
We should support alternate prefixes in version tags. At the moment the
project assumes that the prefix v is always used. But some projects use
no prefix, or include the package as a prefix to the version (e.g. in
monorepos).

All public-facing functions now support a tag prefix option. It is
optional, so this is a non-breaking change. The `updateChangelog`
method accepts an array of tag prefixes instead of a single one,
so that we can better support the first release of a monorepo where
you might want to compare against a backup tag if the first isn't
found.

Fixes #116