Update TypeScript to the latest stable version (#124)
* Update TypeScript to the latest stable version

TypeScript has been updated to the latest stable version. Additionally,
the version range has been updated to use `~` because TypeScript does
not follow SemVer (they make breaking changes as minor updates).

The configuration has been updated to match the module template, except
that two options (`exactOptionalPropertyTypes` and
`noUncheckedIndexedAccess`) have been omitted temporarily to reduce the
number of changes required for this update. They can be added in later
PRs.

Just as in the module template, a separate `tsconfig` file has been
created to distinguish build settings from settings used to compile
unit tests.

The one configuration difference is the compilation target. The target
is ES2019 here because this is a developer tool, so we don't need to
stick to ES2017 for browser compatibility.

[1]: https://devblogs.microsoft.com/typescript/announcing-typescript-4-8/

* Fix handling of SyntaxError in CLI