Allow prerelease versions in release headers (#130)
Currently, this tool expects the version in a release header to be a
version that looks like `<major>.<minor>.<patch>` and rejects other
formats. However, it may be useful to publish a prerelease version of a
package so that integration with other products/services can be tested
in a more "live" manner, particularly if that package is a part of a
monorepo. A prerelease version looks like
`<major>.<minor>.<patch>-<prerelease identifier>.<number>`, where
`<prerelease identifier>` can be an arbitrary word like `alpha`, `beta`,
or `rc`. This commit allows such versions.