Remove `as` assertions, or fix to mitigate the risk of silent failure (#158)
* Replace `as T[]` assertions with `?? []`

- `as` assertions intended to exclude an empty/nullable type can be replaced by a nullish coalescing operator converting nullable values into an acceptable empty value that doesn't pollute the variable's type signature.
- TODO: document and expound in https://github.com/MetaMask/contributor-docs/issues/47

* Replace `as` assertions with redundant type guards

Sometimes TypeScript's type narrowing and inference doesn't fully work quite like it should. Even in these cases, using `as` assertions should be avoided.

It may seem like we know for certain what the type should be and there's no risk of it breaking, but there's always the possibility that some combination of changes elsewhere might affect the typing, in which case `as` will suppress the alerts TypeScript would otherwise provide us with. This may lead to code breaking silently.

In this case, adding type guards and null checks -- even if they're redundant -- is preferable to the above scenario.

- TODO: document and expound in https://github.com/MetaMask/contributor-docs/issues/47

* Make `as` assertion rely on inference instead of hard-coded type casting

If anything changes in the typing of `ChangeCategory` or `unreleasedChanges` that make them inconsistent, this code will fail silently. Relying on type inference instead of explicit type casting will make these line a little less brittle.

- TODO: document and expound in https://github.com/MetaMask/contributor-docs/issues/47

* Fix type narrowing for `currentVersion` by consolidating conditional branches

- more consistent logic for `isReleaseCandidate` null check
- wasted processing overhead due to errors throwing later than strictly necessary

* Install `@metamask/utils` as a devDep

* Replace `Object.keys` with `getKnownPropertyNames` method

- removes need for type casting while iterating over object keys

* Update src/update-changelog.ts