Avoid circular dependency on @metamask/utils (#170)
`@metamask/utils` has a development dependency on this package (since we
need a way to validate its changelog). So, if this package depends on
`@metamask/utils`, then that creates a circular dependency which could
potentially cause issues. Since we only use one function from this
package, we can just copy it.